Get compiling

+++ Copter/OribtBoosterPack/OrbitOled.c
@@ -439,7 +439,7 @@ void
 OrbitOledPutBuffer(int cb, char * rgbTx)
 	{
 	int				ib;
-	unsigned long	bTmp;
+	uint32_t	bTmp;
 
 	/* Bring the slave select line low
 	*/
@@ -488,7 +488,7 @@ OrbitOledPutBuffer(int cb, char * rgbTx)
 char
 Ssi3PutByte(char bVal)
 	{
-	unsigned long	bRx;
+	uint32_t	bRx;
 
 	/* Bring the slave select line low
 	*/
